Vervangen van het WOTD commando
