+++ routes/console.php
@@ -1,11 +1,11 @@
 <?php
 
+use App\Console\Commands\PublishWordOfTheDayOnDiscord;
 use App\Console\Commands\Reminders\InactivityWarningCommand;
-use App\Console\Commands\WordOfTheDayCommand;
 use Illuminate\Support\Facades\Schedule;
 
 // Own custom cron commands
-Schedule::command(WordOfTheDayCommand::class)->daily()->at('00:01');
+Schedule::command(PublishWordOfTheDayOnDiscord::class)->daily()->at('00:01');
 
 // 3th party cron commands
 Schedule::command('ban:delete-expired')->everyMinute();
